+++ addon/utils/registration/selected-option.js
@@ -3,16 +3,17 @@ import { A as emberA } from '@ember/array';
 import { set } from '@ember/object';
 
 export function initSelectedOptions(selectBox) {
+  set(selectBox, '_selectedOptions', emberA());
   set(selectBox, 'selectedOptions', emberA());
 }
 
 export function registerSelectedOption(selectBox, selectedOption) {
-  selectBox.selectedOptions.addObject(selectedOption);
+  selectBox._selectedOptions.addObject(selectedOption);
   scheduleUpdateSelectedOptions(selectBox);
 }
 
 export function deregisterSelectedOption(selectBox, selectedOption) {
-  selectBox.selectedOptions.removeObject(selectedOption);
+  selectBox._selectedOptions.removeObject(selectedOption);
   scheduleUpdateSelectedOptions(selectBox);
 }
 
@@ -21,9 +22,27 @@ function scheduleUpdateSelectedOptions(selectBox) {
 }
 
 function updateSelectedOptions(selectBox) {
+  if (selectBox.isDestroyed || !selectBox.domElement) {
+    return;
+  }
+
+  setSelectedOptions(selectBox, selectBox._selectedOptions);
+}
+
+function setSelectedOptions(selectBox, selectedOptions) {
+  const elements = [
+    ...selectBox.domElement.querySelectorAll(
+      '[data-component="selected-option"]'
+    )
+  ];
+
+  const sort = (a, b) => {
+    return elements.indexOf(a.domElement) - elements.indexOf(b.domElement);
+  };
+
   set(
     selectBox,
     'selectedOptions',
-    emberA(selectBox.selectedOptions.toArray())
+    emberA(selectedOptions.toArray().sort(sort))
   );
 }